requirements.txt
allure-pytest
pytest
pytest-rerunfailures
selenium
